PointerIcon: Create bitmap using density-aware constructor
Instead of manually setting the density of a Bitmap that we create
(which looks like we're working around Bitmap's functionality), use
Bitmap's density-aware constructor that takes a DisplayMetrics object.
It essentially provides the same functionality as setting the density,
but since it's wrapped up in a supported constructor, it increases
readability and removes the need for us to explain what we're doing.

Bug: 305193969
Test: Presubmit
Change-Id: I2b4bdc60e2744a89c8154bd9359de6254a810426