Make AnrTimerTest.testDumpOutput hermetic
testDumpOuput() calls the AnrTimer dumpsys routine and verifies that
the expected timers are present, or not.  It was originally written on
the assumption that the only timers in the process would be those
created in AnrTimerTest, but this is not true: other ActivityManager
tests also create timers that are immortal and thus appear in the
dumpsys output.

This change gives unique names to the timers created in AnrTimerTest
and checks that those timers are present, or not.

Tested by manually modifying the code to inject a random timer name:
verified that the test still produced the correct result.

Test: atest
 * FrameworksServicesTests

Bug: 326043049
Change-Id: I93ea43236b4f9f5f2f9c2e5bd7276459d9bee3d2