[flexiglass] Change to Lockscreen scene when legacy system says to.
StatusBarKeyguardViewManager.show(...) calls
SceneInteractor.changeScene(Lockscreen), allowing the legacy system to
show the lockscreen scene when it would otherwise show the keyguard.

This change was made in support for the attached bug, to make sure that
we show the lockscreen scene when told to by KeyguardService (through
KeyguardViewMediator) because we folded the device to its closed
position.

While it would be much cleaner to have Flexiglass automatically switch
to the Lockscreen scene when told to by KeyguardService (and/or
KeyguardViewMediator), there are many side-effects along the call path
that ends in StatusBarKeyguardViewManager.show (and in that method and
in hide also) that mean that hooking it up properly is not
straightforward. b/328109797 was created to clean this up later.

Fix: 324646158
Test: manually verified that folding the device to its closed position
correctly shows or doesn't show the lockscreen scene based on the
current setting in Settings > Display > Continue using apps on fold and
based on whether the device was previously unlocked or not.
Flag: ACONFIG com.android.systemui.scene_container DEVELOPMENT

Change-Id: I5a9b2cf319636f192750731a0106a4f4d48fedf3