Properly check for malformed extras
In ControlsRequestReceiver, we trust the extras in the intent, but
because it could be assembled by a malicious app, those could be
malformed, causing SystemUI to crash.

Make sure that we properly check for exceptions in loading the
parcellables as well as missing data and return early (after logging an
error).

Fixes: 329369938
Flag: None
Test: manual, using test apk
Test: atest ControlsRequestReceiverTest
Change-Id: I945d2c5a8daf56c0708de5ccdd6020cac857827a