Set screen lock for the parent profile instead of system user
Here we change the logic to call the set screen lock prompt to ensure it
is called from the parent user of the private profile thereby moving
away from the assumption that screen lock needs to be set only on system
users to access private space.

Bug: 332482633
Test: Tested by flashing cheetah_hsum-staging-userdebug and trying to
unlock PS both with and without the screen lock being set.

Change-Id: Id9df5174ad87c25e36bf92efb38e614cdc3d905b