Accept MODE_DEFAULT for ACCESS_NOTIFICATIONS
In commit 8dfdb4f9358e55b7c06004b3af58951607c16613, the mode for
ACCESS_NOTIFICATIONS was changed from MODE_ALLOWED to MODE_DEFAULT. It
means that when noteOp is called, it may now return MODE_DEFAULT.

Change the check on the return value to also accept MODE_DEFAULT. The
actual permission check is done earlier in the three methods, with a
call to the _enforcePermission method.

Also clean up the @EnforcePermission annotations to clarify their scope.
This is a purely esthetic change.

Bug: 332787465
Test: boot & check Notification History
Test: confirm b/329409825 remains fixed
Change-Id: I75a83cc42ab265ad381254c7ecafeb2088e665cc