Check KeyguardTransitionInteractor for the current keyguard state
Don't check KeyguardBypassController or the AlternateBouncerInteractor
for whether the alternate bouncer is visible. It's possible that we
just started transitioning away from the alternate bouncer, but those
states are already reporting isAltBouncerShowing=false. Instead, it's
more robust to check whether we're currently in the KeyguardTransition
state for the alternate bouncer which will include when we're
transitioning away from the AlternateBouncer but haven't finished into
the next state (for example, GONE).

Test: atest BiometricsUnlockControllerTest
Fixes: 332885570
Flag: None
Change-Id: Iea1496f6ba4be62eef57ba2040b7b43c7e730126