Fix wrongly applied material theme for StandardMenuPopup (2nd attempt)
The second attempt contains followings
- Add feature flag for fixing this issue.
- Tweak the layout for not to be misaligned in some case.

Original comment from the first attempt:

The candidate windows has updated for Material theme but
it was done only for CascadingMenuPopup. The same things
need to be done for StandardMenuPopup which is mainly
used for the phone devices.

This CL also fixes the icon position wrongly aligned
to the left of the context menu for some case.

Bug: 332542108
Bug: 334822606
Test: Manually tested with Pixel 8 and Pixel Table on following conditions
- Google Map send feedback menu
- Copy menu on settings in Android version page
- Regular EditText
Change-Id: I181fb046e74033c94657c5455e8b70ea6efa159e