Use NDK version of transaction completed callback
The addTransactionCompletedListener doesn't work
on 24D1-dev branch, it requires API flag that
is enabled only in V.

There is an NDK version of the same method
that is available since API 29 but it is not
exposed to Java. Replaced Java call to
addTransactionCompletedListener with an NDK
call to ASurfaceTransaction_setOnComplete.

Bug: 334873960
Test: fold/unfold without delay on 24D1-dev
Test: atest UnfoldToLauncherMicrobenchmark
Merged-In: Ibed2cf715cc3898167cab02e21405bf028ca1111
Change-Id: Id6dfd4615343ed524582eb62aa24dd3e0a1413b5