Add PipTransitionState into pip2
Create a new PipTransitionState class
to dispatch state updates for
PiP specifically guided by transitions.

Also connect scheduling expand-pip to onActivityRestartAttempt().
But remove the usages of onActivityPinned/Unpinned() since we can now
rely on PipTraansitionStateChanged listener updates instead which
are in sync with diffferent stages of running Transitions.

The plan is to use these state changes to enable and disable touches
in pip2.PipTouchHandler in a follow up CL (a substitute for onAct(Un)Pinned).

Bug: 336870419
Test: swipe-up to enter PiP then tap icon to expand
Change-Id: Ica297920fbf4ace1c18b17e92341e179dcce22d3