Add a finalizer to LockscreenCredential to wipe the LSKF on GC
Adds a finalizer to the `LockscreenCredential` class, which calls
the `zeroize` method to zeroize the LSKF `byte[]`. This ensures that
`LockscreenCredential` objects that were not closed still have their
data wiped when they are garbage collected.

This is part of an effort to remove instances of the LSKF available
in a RAMdump.

Bug: 320392352
Test: atest com.android.server.locksettings
Change-Id: Id7f1f102cdaa182659e2632bdaf81bd2c46d0f86